Se crean entities.

+++ src/main/java/com/platzi/pizza/persitence/entity/OrderItemEntity.java
@@ -18,14 +18,14 @@ import lombok.Setter;
 public class OrderItemEntity {
     @Id
     @Column(name = "id_order", nullable = false)
-    private Long idOrder;
+    private Integer idOrder;
 
     @Id
     @Column(name = "id_item", nullable = false)
-    private Long idItem;
+    private Integer idItem;
 
     @Column(name = "id_pizza", nullable = false)
-    private Long idPizza;
+    private Integer idPizza;
 
     @Column(nullable = false, columnDefinition = "Decimal(2,1)")
     private Double quantity;
